[refactor] modal 경로 수정
